Add code to watch MPD and play next song,.
Have the Dancer app fork a new process to run this code upon startup.

(In the old Catalyst-based setup this is adapted from, it was a separate process
which had to be started manually.)

Might need to add some code to monitor whether the other process is still
running, and start another one if it dies.